Move service account to its own module
